Fixes issue where modules wouldn't load
Without this option, even though the config_dir is parsed in the run command, modules will fail to load because they cannot be found, presumably because node is looking in the OS root instead of the actual folder. CD'ing into the directory before executing fixes the problem, at least for me. Now I'm a happy camper. :)