Background node process only
 Cleans up stop_process and saves memory. Git doesn't take kindly to long one liners, hence the split.

One potential issue is if this is in anyone's poorly implemented CI, the update will result in two node instances, one working zombie and one probably not. To have this auto update properly, you should kill with the current version and start with the new one... not update then restart. I don't think that's too much of an issue as anyone who seriously uses this has probably customized it... TLS ports, directories etc